pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "booty"
version = "0.1.0"
description = "GitHub issue to PR agent pipeline"
requires-python = ">=3.11"
dependencies = [
    "click",
    "fastapi[standard]",
    "pydantic-settings",
    "gitpython",
    "structlog",
    "asgi-correlation-id",
    "uvicorn",
    "magentic[anthropic]",
    "PyGithub",
    "anthropic",
    "pathspec",
    "pyyaml",
    "tenacity",
    "giturlparse",
    "sentry-sdk",
    "pip-audit",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[project.scripts]
booty = "booty.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
